# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Synie API
  version: 0.1.0
  description: Synie ERP Go 主 API。金额与小数在 JSON 中一律使用字符串。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /base/accounts/query:
    post:
      operationId: queryBasAccounts
      tags: [base]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ListQuery" } } } }
      responses:
        "200": { description: 会计科目分页结果, content: { application/json: { schema: { $ref: "#/components/schemas/AccountList" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /base/accounts:
    post:
      operationId: createBasAccount
      tags: [base]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AccountCreate" } } } }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /base/accounts/init-template:
    post:
      operationId: initializeBasAccountsTemplate
      tags: [base]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AccountTemplateInitialize" } } } }
      responses:
        "201": { description: 模板已初始化, content: { application/json: { schema: { $ref: "#/components/schemas/AccountTemplateResult" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /base/accounts/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      operationId: getBasAccount
      tags: [base]
      responses:
        "200": { description: 会计科目, content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateBasAccount
      tags: [base]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AccountUpdate" } } } }
      responses:
        "200": { description: 已更新, content: { application/json: { schema: { $ref: "#/components/schemas/Account" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteBasAccount
      tags: [base]
      responses:
        "204": { description: 已删除 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [unauthorized, rate_limited, forbidden, validation, not_found, conflict, not_implemented, internal]
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ListQuery:
      type: object
      additionalProperties: false
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 200
          default: 20
        offset:
          type: integer
          minimum: 0
          default: 0
        search:
          type: string
          maxLength: 256
        sort:
          $ref: "#/components/schemas/Sort"
        filter:
          $ref: "#/components/schemas/FilterState"
    Sort:
      type: object
      additionalProperties: false
      required: [column, direction]
      properties:
        column:
          type: string
        direction:
          type: string
          enum: [ascending, descending]
    FilterState:
      type: object
      additionalProperties:
        type: object
        additionalProperties: true
        required: [kind]
        properties:
          kind:
            type: string
            enum: [text, bool, enum, enumArray, number, date, fk, polyFk]
...
    AccountDirection:
      type: string
      enum: [DEBIT, CREDIT]
    AccountRole:
      type: string
      enum:
        - UNBILLED_RECEIVABLE
        - RECEIVABLE
        - ADVANCE_RECEIVED
        - UNBILLED_PAYABLE
        - PAYABLE
        - OTHER_PAYABLE
        - ADVANCE_PAID
        - TRAVEL
        - OFFICE
        - ENTERTAINMENT
        - TRANSPORT
        - OTHER_EXPENSE
    AccountReference:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    Account:
      type: object
      required: [id, code, name, direction, isGroup, active, companyId, company, hasChildren, insertedAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        direction: { $ref: "#/components/schemas/AccountDirection" }
        isGroup: { type: boolean }
        active: { type: boolean }
        role:
          allOf: [{ $ref: "#/components/schemas/AccountRole" }]
          nullable: true
        parentId: { type: string, format: uuid, nullable: true }
        companyId: { type: string, format: uuid }
        currencyId: { type: string, format: uuid, nullable: true }
        parent:
          allOf: [{ $ref: "#/components/schemas/AccountReference" }]
          nullable: true
        company: { $ref: "#/components/schemas/AccountReference" }
        currency:
          allOf: [{ $ref: "#/components/schemas/AccountReference" }]
          nullable: true
        hasChildren: { type: boolean }
        insertedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AccountCreate:
      type: object
      additionalProperties: false
      required: [code, name, direction, companyId]
      properties:
        code: { type: string, minLength: 1, maxLength: 32 }
        name: { type: string, minLength: 1, maxLength: 128 }
        direction: { $ref: "#/components/schemas/AccountDirection" }
        isGroup: { type: boolean, default: false }
        active: { type: boolean, default: true }
        role:
          allOf: [{ $ref: "#/components/schemas/AccountRole" }]
          nullable: true
        parentId: { type: string, format: uuid, nullable: true }
        companyId: { type: string, format: uuid }
        currencyId: { type: string, format: uuid, nullable: true }
    AccountUpdate:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        direction: { $ref: "#/components/schemas/AccountDirection" }
        isGroup: { type: boolean }
        active: { type: boolean }
        role:
          allOf: [{ $ref: "#/components/schemas/AccountRole" }]
          nullable: true
        parentId: { type: string, format: uuid, nullable: true }
        currencyId: { type: string, format: uuid, nullable: true }
    AccountList:
      type: object
      required: [count, results]
      properties:
        count: { type: integer, format: int64 }
        results: { type: array, items: { $ref: "#/components/schemas/Account" } }
    AccountTemplateInitialize:
      type: object
      additionalProperties: false
      required: [companyId, template]
      properties:
        companyId: { type: string, format: uuid }
        template:
          type: string
          enum: [CAS, SMALL, INTL]
    AccountTemplateResult:
      type: object
      required: [createdCount]
      properties:
        createdCount: { type: integer }